Refactoring
* Move OSUtils class to the "files" package, where it belongs
* Move "id" parameter closer to the end of path. This allows to use
  REST client and resource for test cases, where "id" field is not required
* Temporary filenames are more consistent
