+++ http/rest-client-reactive/src/main/java/io/quarkus/ts/http/restclient/reactive/files/OsUtils.java
@@ -1,4 +1,4 @@
-package io.quarkus.ts.http.restclient.reactive;
+package io.quarkus.ts.http.restclient.reactive.files;
 
 import java.io.IOException;
 import java.io.InputStream;
