+++ http/rest-client-reactive/src/main/java/io/quarkus/ts/http/restclient/reactive/json/IdBeanParam.java
@@ -1,4 +1,4 @@
-package io.quarkus.ts.http.restclient.reactive;
+package io.quarkus.ts.http.restclient.reactive.json;
 
 import javax.ws.rs.PathParam;
