+++ http/rest-client-reactive/src/main/resources/classic.properties
@@ -1,1 +1,1 @@
-io.quarkus.ts.http.restclient.reactive.ReactiveRestInterface/mp-rest/url=http://localhost:${quarkus.http.port}
+io.quarkus.ts.http.restclient.reactive.json.JsonRestInterface/mp-rest/url=http://localhost:${quarkus.http.port}
