+++ http/rest-client-reactive/src/main/resources/modern.properties
@@ -1,5 +1,5 @@
 # new schema introduced in https://github.com/quarkusio/quarkus/pull/17220 . See classic.properties for old-style settings
-quarkus.rest-client."io.quarkus.ts.http.restclient.reactive.ReactiveRestInterface".url=http://localhost:${quarkus.http.port}
+quarkus.rest-client."io.quarkus.ts.http.restclient.reactive.json.JsonRestInterface".url=http://localhost:${quarkus.http.port}
 quarkus.rest-client."io.quarkus.ts.http.restclient.reactive.files.FileClient".url=http://localhost:${quarkus.http.port}
 quarkus.http.limits.max-body-size=3G
 quarkus.rest-client.read-timeout = 60000
